made Delivery immutable. Created Communcateable unit to get replies for the truck intention. Simplified TruckScheduleUnit and also made it immutable. TimeSlot is still not immutable, which it should be.  Please enter the commit message for your changes. Lines starting

+++ src/main/java/shaz/rmc/core/domain/Order.java
@@ -84,7 +84,7 @@ public class Order implements Serializable {
         return requiredDischargePerHour;
     }
     public int getRequiredDischargePerMinute(){
-    	return (int)Math.ceil(requiredDischargePerHour/60);
+    	return (int)Math.ceil((double)requiredDischargePerHour/60d);
     }
     
     public boolean hasRequiredDischargeRate() {
